+++ src/main/kotlin/com/github/strindberg/emacsj/actions/kill/AppendNextKillAction.kt
@@ -1,0 +1,6 @@
+package com.github.strindberg.emacsj.actions.kill
+
+import com.github.strindberg.emacsj.kill.AppendNextKillHandler
+import com.intellij.openapi.editor.actions.TextComponentEditorAction
+
+class AppendNextKillAction : TextComponentEditorAction(AppendNextKillHandler())