+++ src/main/kotlin/com/github/strindberg/emacsj/actions/kill/CopyRegionAction.kt
@@ -1,0 +1,6 @@
+package com.github.strindberg.emacsj.actions.kill
+
+import com.github.strindberg.emacsj.kill.CopyRegionHandler
+import com.intellij.openapi.editor.actions.TextComponentEditorAction
+
+class CopyRegionAction : TextComponentEditorAction(CopyRegionHandler())