+++ src/main/kotlin/com/github/strindberg/emacsj/actions/kill/CutRegionAction.kt
@@ -1,0 +1,6 @@
+package com.github.strindberg.emacsj.actions.kill
+
+import com.github.strindberg.emacsj.kill.CutRegionHandler
+import com.intellij.openapi.editor.actions.TextComponentEditorAction
+
+class CutRegionAction : TextComponentEditorAction(CutRegionHandler())